requirements.txt
asgiref==3.10.0
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg==3.2.12
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
